+++ Assets/00.Scenes/00.TestScene/SpawnTest.unity
@@ -565,7 +565,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 1749898507}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 6ac36e53116b4ad4c8769829ee3eeb6c, type: 3}
+  m_Script: {fileID: 11500000, guid: 081a58e8bb9aedf49badfaf2c0d45ec2, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   friendlyPrefab: {fileID: 7654564522672126550, guid: 44d81176aa2a88040bc6d878b02038bb, type: 3}
