--- Assets/01.Scripts/02.Unit/UnitSpawner2.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 6ac36e53116b4ad4c8769829ee3eeb6c
