+++ ProjectSettings/EditorBuildSettings.asset
@@ -21,5 +21,6 @@ EditorBuildSettings:
     path: Assets/00.Scenes/00.TestScene/SpawnTest.unity
     guid: 963c4ea7f9b49f54d931142c13d9eb1d
   m_configObjects:
+    com.unity.addressableassets: {fileID: 11400000, guid: 07ab3b770b9b1ac4799faf6404832ae6, type: 2}
     com.unity.input.settings.actions: {fileID: -944628639613478452, guid: 2bcd2660ca9b64942af0de543d8d7100, type: 3}
   m_UseUCBPForAssetBundles: 0
